+++ lucid/models/__init__.py
@@ -17,7 +17,7 @@ from lucid.models._base   import ModelConfig, PretrainedModel
 from lucid.models._meta   import ModelFamilyMeta, model_family_meta
 from lucid.models._mixins import (
     BackboneMixin, ClassificationHeadMixin, DiffusionMixin,
-    FeatureInfo, GenerationMixin, MaskedLMMixin,
+    FeatureInfo, CausalLMMixin, MaskedLMMixin,
 )
 from lucid.models._output import (
     ModelOutput, BaseModelOutput, BaseModelOutputWithPooling,
@@ -408,7 +408,7 @@ __all__ = [
     "ModelConfig", "PretrainedModel",
     "ModelFamilyMeta", "model_family_meta",
     "BackboneMixin", "ClassificationHeadMixin", "DiffusionMixin",
-    "FeatureInfo", "GenerationMixin", "MaskedLMMixin",
+    "FeatureInfo", "CausalLMMixin", "MaskedLMMixin",
     "LanguageModelConfig", "TextActivation",
     "GenerativeModelConfig", "DiffusionModelConfig",
     "GenerativeActivation", "BetaSchedule",